+++ tests/common/mod.rs
@@ -49,12 +49,17 @@ impl BitFlags {
 
 #[allow(unused)]
 pub fn new_1008(transactions: &[I2cTrans]) -> Kxcj9<I2cMock, ic::Kxcj9_1008> {
-    Kxcj9::new_1008(I2cMock::new(&transactions), SlaveAddr::default())
+    Kxcj9::new_kxcj9_1008(I2cMock::new(&transactions), SlaveAddr::default())
 }
 
 #[allow(unused)]
 pub fn new_1018(transactions: &[I2cTrans]) -> Kxcj9<I2cMock, ic::Kxcj9_1018> {
-    Kxcj9::new_1018(I2cMock::new(&transactions), SlaveAddr::default())
+    Kxcj9::new_kxcj9_1018(I2cMock::new(&transactions), SlaveAddr::default())
+}
+
+#[allow(unused)]
+pub fn new_1041(transactions: &[I2cTrans]) -> Kxcj9<I2cMock, ic::Kxcj9_1008> {
+    Kxcj9::new_kxcjb_1041(I2cMock::new(&transactions), SlaveAddr::default())
 }
 
 pub fn destroy<IC>(sensor: Kxcj9<I2cMock, IC>) {